KAN-27-fix:swagger 비인증 url 추가
fix:swagger 비인증 url 추가

swagger 비인증 url 추가

+++ announcement/src/main/java/org/pda/announcement/stockprice/domain/Stockpricemonth.java
@@ -48,7 +48,7 @@ public class Stockpricemonth {
     private Integer closePrice;
 
     @Column(name = "volume")
-    private Integer volume;
+    private Long volume;
 
     @Column(name = "change_rate")
     private Float changeRate;
